1、商品相关筛选条件（抢购超时下架） 2、抢购降价逻辑bug修复 3、根据阿里代码检测修改部分不规范代码 4、修改了用户判断逻辑，列表中用户token失效时不做验证

+++ fangyuanyouyue-microservice/fangyuanyouyue-microservice-user/src/main/java/com/fangyuanyouyue/user/service/impl/UserInfoServiceImpl.java
@@ -303,7 +303,7 @@ public class UserInfoServiceImpl implements UserInfoService {
             UserThirdParty userThirdParty = userThirdPartyMapper.getUserByThirdNoType(unionId,type);
             //校验是否绑定三方账号
             if(userThirdParty != null){
-                if(userThirdParty.getUserId() == userInfo.getId()){
+                if(userThirdParty.getUserId().intValue() == userInfo.getId().intValue()){
                     throw new ServiceException("请勿重复绑定！");
                 }else{
                     UserInfo user = userInfoMapper.selectByPrimaryKey(userThirdParty.getUserId());
@@ -717,7 +717,7 @@ public class UserInfoServiceImpl implements UserInfoService {
 	public void registIMUser(UserInfo user) throws ServiceException {
         try {
             //判断用户是否已经注册环信
-            if(user.getIsRegHx()==null||user.getIsRegHx()==StatusEnum.NO.getCode()) {
+            if(user.getIsRegHx()==null||user.getIsRegHx().intValue()==StatusEnum.NO.getCode().intValue()) {
                 schedualMessageService.easemobRegist(user.getId().toString(), MD5Util.MD5("xiaofangyuan"+user.getId().toString()));
                 user.setIsRegHx(StatusEnum.YES.getCode());
                 userInfoMapper.updateByPrimaryKey(user);
